Update SlackPirate.py
Made a couple changes - uncommented the dump_user_list function as that was my mistake I committed last night.

Also removed the UA header from the --cookie flag function - reason being that the format and response of the page will be different depending on the UA provided to the page which means it won't work a lot of the time. Saying that, I don't think it really matters if we don't have a UA header here as these won't show up in the Workspace audit logs.